Moved tests to `test__graph_input_check.py`
